refactor batchUpsert code to standalone trait
